FullForm & OutputForm should not use MathML
Put formatting decisions more in control of mathics-django.

To do this we set the output format to be "unformatted". This alows
format.py to make decisions about the final formatting.

In particular, we want OutpuForm and FullForm *not* to produce MathML.
Instead it should produce plain-ol' text which you can cut and paste.

Worse, OutputForm[Rectangle[]] was producing a graphic when it shouldn't
have been.

There still seems to be a bug with FullForm not expanding graphics
primitives, but that can come later.

Note that we also handle matplotlib filenames better now as well since
that is no longer MathML.