Fold in some improvements from Mathics Core..
This gets us back some of the test data in examples but we still
need to do the sections inside guide sections.

I think from here it will be an easy reach.

However all of this code needs DRY'd more with mathics core code, pulled
out of here and put in a module.

And even, totally rewritten!