exceptions are formatted better now...
Better handling and formatting of Syntax Errors and Python Exceptions

views.py:
--------

We now use the "form" parameter to detect syntax errors and to mark exception errors that are to be displayed.

This requires a recent Mathics 6.0.0

mathics.js:
-----------

Uses the tag set up for it in views.py to it to detect "Syntax Error" and "Python Exception" and formats these in a custom way that better displays these.

More work could be done for Python Exception in the area of custom traceback rendering. For example, file paths could be incorporated in link () tags.

A previous commit reformatted the file to remove tabbing and to make explicit (for Emacs) indent of 4 spaces.